DAD-68 & DAD-72 | modifying the ui to update system config, and added input fields for sparse embedding options
